requirements_txt.txt
# Core ML/DL
torch>=2.0.0
torch-geometric>=2.3.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Bayesian & Causal
pymc>=5.7.0
arviz>=0.15.0
dowhy>=0.9.0
causalml>=0.15.0

# RL & Simulation
gymnasium>=0.29.0
stable-baselines3>=2.0.0
mesa>=2.1.0

# Database & ETL
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# API & Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
aiohttp>=3.8.0

# Time Series & Stats
statsmodels>=0.14.0
prophet>=1.1.0

# Visualization & UI
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
dash>=2.14.0

# Portfolio Optimization
cvxpy>=1.4.0
pypfopt>=1.5.0

# Explainability
shap>=0.42.0
lime>=0.2.0

# AI Chatbot
openai>=1.3.0

# Config & Environment
pydantic>=2.4.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Testing & Quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0

# Monitoring & Alerts
schedule>=1.2.0
telegram-bot>=20.0.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
dask>=2023.9.0
redis>=5.0.0